requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
chardet==4.0.0
click==8.1.8
dataclasses==0.6
deprecation==2.1.0
fastapi==0.115.6
gotrue==0.2.0
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
postgrest-py==0.4.0
py==1.11.0
pydantic==2.10.4
pydantic_core==2.27.2
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
supabase-py==0.0.2
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
websockets==9.1
